Move all SDL calling code to _sdl_utils
